<master> Remove SPV, Add IAddressmanager interface to mock it in tests

+++ NBitcoin/JsonConverters/FeeRateJsonConverter.cs
@@ -25,7 +25,7 @@ namespace NBitcoin.JsonConverters
 			try
 			{
 				if(reader.TokenType == JsonToken.Integer)
-					return new FeeRate((long)reader.Value);
+					return new FeeRate((decimal)reader.Value);
 				if(reader.TokenType == JsonToken.Float)
 					return new FeeRate((decimal)(double)reader.Value);
 				throw new JsonObjectException("Fee rate amount should be in satoshi", reader);
